refactor(cli): extract auth routes into server-routes/auth.ts
#354 pass 5. Now that the auth-session is a clean dependency (server-auth.ts),
move the 5 auth endpoints (/api/auth/status, get-session, logout, sign-out, and
the CLI OAuth login flow with its temporary callback server) into
registerAuthRoutes(app, deps), following the established route-module pattern.

deps: { cloudApiBaseUrl, readLocalAuthSession, isAuthValid, clearLocalAuthSession,
autoSyncInsights }. saveAuthToken is imported directly by the module and dropped
from server.ts's imports.

server.ts: 3097 -> 2967 lines; new auth.ts is 157. Verbatim move, no behavior
change — 340 CLI tests pass, typecheck + lint clean, server boot + auth wiring
verified via editor-server + cli-smoke e2e.

Refs #354
